Create exome_auto_gatk_v2.sh
GATK based exome auto pipeline from fastq to vcf without control samples. Mutation calling on a singe sample. This can father be exploited to run  gatk pre - processing on individual samples and then parse them in mutect2 or varscan2 caller for germline and somatic mutation calling. As of now it is build on hg19 genome and gatk 3.5 with bundle 2.8